Switch piccolo and hit colors
Piccolo was set as purplish-blue, and hit was set as green. I assumed this was a mistake, so I fixed it.

I know other colors don't match. For example, goten being white for no apparent reason. I'm willing to accept that. But piccolo and hit being purple and green side to side here in the site is too blatant and painful to bear: https://moon.io/colours